Commit 8
Added a large 8x14 grid of squares, each coloured white (#FFFFFF) with the text "0" in the middle of them. This is the gameplay area where the breakable blocks will be. Also rearranged the rightPane to accommodate the top of the fire button aligning with the bottom of the gameplay area.