Commit 11
Random Number Generator generates ArrayList of type integer to be used to fill the blocks text with random integers between 20 and 50 inclusive.
Added a test button that will allow me to advance to the next round upon clicking. Button removes all rectangles and text from the stackPanes, and removes all stackPanes from the floatPane.

Problem currently resides in the shifting-down of rows; it doesn't shift down. A new row is simply added to the bottom instead of being pushed onto the top.